GS-137: Renaming files and folders for consistancy.
