+++ 01-atoms/buttons/Button.stories.js
@@ -1,5 +1,4 @@
 import React from 'react';
-
 import Button from './Button.component';
 
 /**
