+++ 01-atoms/images/Figure/figure.stories.js
