+++ 01-atoms/images/Logo/Logo.component.js
