+++ 01-atoms/images/Picture/Picture.stories.js
