+++ 01-atoms/images/ResponsiveImage/ResponsiveImage.stories.js
