+++ 01-atoms/images/icons/FacebookIcon.component.js
