+++ 01-atoms/images/icons/MenuIcon.component.js
