+++ 01-atoms/links/Link.stories.js
@@ -1,5 +1,4 @@
 import React from 'react';
-
 import Link from './Link.component';
 
 /**
