+++ 02-molecules/TextWithImage/TextWithImage.component.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import ResponsiveImage from '../../01-atoms/image/ResponsiveImage.component';
-import { imageTypeProps } from '../../01-atoms/image/Image.component';
+import ResponsiveImage from '../../01-atoms/images/ResponsiveImage/ResponsiveImage.component';
+import { imageTypeProps } from '../../01-atoms/images/Image/Image.component';
 import Paragraph from '../../01-atoms/text/Paragraph.component';
 import bem from '../../_utils/bem';
