+++ 03-organisms/site/SiteFooter/SiteFooter.component.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes, { arrayOf } from 'prop-types';
-import InlineMenu from '../../../02-molecules/menus/inline/InlineMenu.component';
-import SocialMenu from '../../../02-molecules/menus/social/SocialMenu.component';
+import InlineMenu from '../../../02-molecules/menus/InlineMenu/InlineMenu.component';
+import SocialMenu from '../../../02-molecules/menus/SocialMenu/SocialMenu.component';
 import { menuItemPropType } from '../../../02-molecules/menus/MenuItem.component';
 import bem from '../../../_utils/bem';
