+++ 03-organisms/site/SiteFooter/SiteFooter.stories.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import SiteFooter from './SiteFooter.component';
-import { items } from '../../../02-molecules/menus/inline/InlineMenu.data.json';
+import { items } from '../../../02-molecules/menus/InlineMenu/InlineMenu.data.json';
 
 /**
  * Storybook Definition.
