+++ 03-organisms/site/SiteHeader/SiteHeader.stories.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import SiteHeader from './SiteHeader.component';
-import { items as breadcrumbs } from '../../../02-molecules/menus/breadcrumbs/breadcrumbs.data.json';
-import { items as menu } from '../../../02-molecules/menus/main/mainmenu.data.json';
+import { items as breadcrumbs } from '../../../02-molecules/menus/Breadcrumbs/Breadcrumbs.data.json';
+import { items as menu } from '../../../02-molecules/menus/MainMenu/MainMenu.data.json';
 
 /**
  * Storybook Definition.
